Move jdk.java.net under vendor Oracle (#17)
The builds from jdk.java.net are Oracle OpenJDK builds produced by Oracle. also, JDK 14 build isn't an EA build; it's GA build of JDK 14.0.2, JDK 15 build is currently RC build and is going to be GA soon-ish. as the download page clearly says the build's type right in its title when you open it, it doesn't make much sense to have build-type in the product name and update it every time it gets changed.